Added saving and loading the game

+++ src/command/Delete.java
@@ -6,12 +6,14 @@ public class Delete extends Command {
     @Override
     public String execute() {
         File file = new File("save.txt");
-        if (file.delete()) {
-            return "You have successfully deleted your save file!";
-        } else {
-            return "The save file couldn't be deleted";
+        if(file.exists()) {
+            if (file.delete()) {
+                return "You have successfully deleted your save file!";
+            } else {
+                return "The save file couldn't be deleted";
+            }
         }
-
+        return "You don't have a save file!";
     }
 
     @Override
